requirements.txt
streamlit
pandas
scikit-learn
requests
fpdf
python-dotenv
google-generativeai
